@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>false</ImplicitUsings>
     <Nullable>enable</Nullable>
     
@@ -15,7 +15,10 @@
 
   <ItemGroup>
     <PackageReference Include="CsvHelper" Version="29.0.0" />
-    <PackageReference Include="Timeline" Version="0.1.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Nuget.Timeline\Timeline\Timeline.csproj" />
   </ItemGroup>
 
 </Project>
