@@ -3,51 +3,50 @@ using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace Finance.JSON
+namespace Finance.Accounting.JSON;
+
+public class ArrayAccountConverter : JsonConverter<Account>
 {
-    public class ArrayAccountConverter : JsonConverter<Account>
+    public override Account Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        public override Account Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        if (reader.TokenType != JsonTokenType.StartArray)
         {
-            if (reader.TokenType != JsonTokenType.StartArray)
-            {
-                throw new JsonException();
-            }
-
-            var parts = new List<string>();
-            while (reader.Read() && reader.TokenType != JsonTokenType.EndArray)
-            {
-                var part = reader.GetString();
-
-                parts.Add(part);
-            }
-
-            return new Account(parts.ToArray());
+            throw new JsonException();
         }
 
-        public override void Write(Utf8JsonWriter writer, Account value, JsonSerializerOptions options)
+        var parts = new List<string>();
+        while (reader.Read() && reader.TokenType != JsonTokenType.EndArray)
         {
-            writer.WriteStartArray();
+            var part = reader.GetString();
 
-            foreach (var level in value.Branch)
-            {
-                writer.WriteStringValue(level);
-            }
-
-            writer.WriteEndArray();
+            parts.Add(part);
         }
 
-        public override Account ReadAsPropertyName(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            var inlineAccount = reader.GetString();
+        return new Account(parts.ToArray());
+    }
 
-            // Technically incorrect as the parts may contain ":"
-            return new Account(inlineAccount.Split(":"));
+    public override void Write(Utf8JsonWriter writer, Account value, JsonSerializerOptions options)
+    {
+        writer.WriteStartArray();
+
+        foreach (var level in value.Branch)
+        {
+            writer.WriteStringValue(level);
         }
 
-        public override void WriteAsPropertyName(Utf8JsonWriter writer, Account value, JsonSerializerOptions options)
-        {
-            writer.WritePropertyName(string.Join(":", value.Branch));
-        }
+        writer.WriteEndArray();
+    }
+
+    public override Account ReadAsPropertyName(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    {
+        var inlineAccount = reader.GetString();
+
+        // Technically incorrect as the parts may contain ":"
+        return new Account(inlineAccount.Split(":"));
+    }
+
+    public override void WriteAsPropertyName(Utf8JsonWriter writer, Account value, JsonSerializerOptions options)
+    {
+        writer.WritePropertyName(string.Join(":", value.Branch));
     }
 }
