@@ -2,64 +2,70 @@ using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace Finance.JSON
+namespace Finance.Accounting.JSON;
+
+internal static class JsonReaderExtensions
 {
-    internal static class JsonReaderExtensions
+    public static void AssertStartArray(this Utf8JsonReader reader)
     {
-        public static void AssertStartArray(this Utf8JsonReader reader)
+        if (reader.TokenType != JsonTokenType.StartArray)
         {
-            if (reader.TokenType != JsonTokenType.StartArray)
-            {
-                throw new JsonException();
-            }
-        }
-
-        public static void AssertEndArray(this Utf8JsonReader reader)
-        {
-            if (reader.TokenType != JsonTokenType.EndArray)
-            {
-                throw new JsonException();
-            }
-        }
-
-        public static void AssertStartObject(this Utf8JsonReader reader)
-        {
-            if (reader.TokenType != JsonTokenType.StartObject)
-            {
-                throw new JsonException();
-            }
-        }
-
-        public static void AssertEndObject(this Utf8JsonReader reader)
-        {
-            if (reader.TokenType != JsonTokenType.EndObject)
-            {
-                throw new JsonException();
-            }
+            throw new JsonException();
         }
     }
 
-    public class DateDateTimeConverter : JsonConverter<DateTime>
+    public static void AssertEndArray(this Utf8JsonReader reader)
     {
-        public override DateTime Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        if (reader.TokenType != JsonTokenType.EndArray)
         {
-            var value = reader.GetDateTime();
-
-            return value;
+            throw new JsonException();
         }
+    }
 
-        public override void Write(Utf8JsonWriter writer, DateTime value, JsonSerializerOptions options)
+    public static void AssertStartObject(this Utf8JsonReader reader)
+    {
+        if (reader.TokenType != JsonTokenType.StartObject)
         {
-            var date = value.ToString("yyyy-MM-dd");
-
-            writer.WriteStringValue(date);
+            throw new JsonException();
         }
+    }
 
-        public override void WriteAsPropertyName(Utf8JsonWriter writer, DateTime value, JsonSerializerOptions options)
+    public static void AssertEndObject(this Utf8JsonReader reader)
+    {
+        if (reader.TokenType != JsonTokenType.EndObject)
         {
-            var date = value.ToString("yyyy-MM-dd");
-
-            writer.WritePropertyName(date);
+            throw new JsonException();
         }
+    }
+}
+
+public class DateDateTimeConverter : JsonConverter<DateTime>
+{
+    public override DateTime Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    {
+        var value = reader.GetDateTime();
+
+        return value;
+    }
+
+    public override DateTime ReadAsPropertyName(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    {
+        var value = reader.GetString();
+
+        return DateTime.Parse(value);
+    }
+
+    public override void Write(Utf8JsonWriter writer, DateTime value, JsonSerializerOptions options)
+    {
+        var date = value.ToString("yyyy-MM-dd");
+
+        writer.WriteStringValue(date);
+    }
+
+    public override void WriteAsPropertyName(Utf8JsonWriter writer, DateTime value, JsonSerializerOptions options)
+    {
+        var date = value.ToString("yyyy-MM-dd");
+
+        writer.WritePropertyName(date);
     }
 }
