@@ -1,110 +1,111 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 using Timeline;
 
-namespace Finance
+namespace Finance.Accounting;
+
+/// <summary>
+/// A mapping from <typeparamref name="Time"/> to transactions.
+/// </summary>
+/// <typeparam name="Time">The type of date or time that points to transactions</typeparam>
+public class Journal<Time> : ITimeline<Time, List<Transaction>>
+    where Time : notnull, IComparable<Time>
 {
-    public class Journal<Time>
-        : ITimeline<Time, List<Transaction>>
-        where Time : notnull, IComparable<Time>
+    private readonly DictionaryTimeline<Time, List<Transaction>> _timeline;
+
+    public Journal()
     {
-        private readonly DictionaryTimeline<Time, List<Transaction>> _timeline;
+        _timeline = new();
+    }
 
-        public Journal()
+    public IEnumerable<List<Transaction>> Values
+    {
+        get
         {
-            _timeline = new();
-        }
-
-        public IEnumerable<List<Transaction>> Values
-        {
-            get
+            foreach (var entry in _timeline)
             {
-                foreach (var entry in _timeline)
-                {
-                    yield return entry.Event;
-                }
+                yield return entry.Event;
             }
         }
+    }
 
-        public List<Transaction> this[Time time]
+    public List<Transaction> this[Time time]
+    {
+        get
         {
-            get
-            {
-                return _timeline[time];
-            }
-            set
-            {
-                _timeline[time] = value;
-            }
+            return _timeline[time];
+        }
+        set
+        {
+            _timeline[time] = value;
+        }
+    }
+
+    public virtual void AddTransaction(Time time, Transaction transaction)
+    {
+        if (!_timeline.HasEvent(time))
+        {
+            _timeline[time] = new();
         }
 
-        public virtual void AddTransaction(Time time, Transaction transaction)
-        {
-            if (!_timeline.HasEvent(time))
-            {
-                _timeline[time] = new();
-            }
+        _timeline[time].Add(transaction);
+    }
 
-            _timeline[time].Add(transaction);
-        }
-
-        public void Add(Journal<Time> journal)
+    public void Add(Journal<Time> journal)
+    {
+        foreach (var entry in journal)
         {
-            foreach (var entry in journal)
+            foreach (var transaction in entry.Event)
             {
-                foreach (var transaction in entry.Event)
-                {
-                    AddTransaction(entry.Time, transaction);
-                }
+                AddTransaction(entry.Time, transaction);
             }
         }
+    }
 
-        public override string? ToString()
+    public override string? ToString()
+    {
+        var builder = new StringBuilder();
+        builder.AppendLine("Journal:");
+
+        foreach (var point in _timeline)
         {
-            var builder = new StringBuilder();
-            builder.AppendLine("Journal:");
+            builder.AppendLine($"{point.Time}");
 
-            foreach (var point in _timeline)
+            foreach (var transaction in point.Event)
             {
-                builder.AppendLine($"{point.Time}");
-
-                foreach (var transaction in point.Event)
-                {
-                    builder.AppendLine(transaction.ToString());
-                }
-
-                builder.AppendLine();
+                builder.AppendLine(transaction.ToString());
             }
 
-            return builder.ToString();
+            builder.AppendLine();
         }
 
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return _timeline.GetEnumerator();
-        }
+        return builder.ToString();
+    }
 
-        public bool HasEvent(Time time)
-        {
-            return _timeline.HasEvent(time);
-        }
+    IEnumerator IEnumerable.GetEnumerator()
+    {
+        return _timeline.GetEnumerator();
+    }
 
-        public IEnumerator<Point<Time, List<Transaction>>> GetEnumerator()
-        {
-            return _timeline.GetEnumerator();
-        }
+    public bool HasEvent(Time time)
+    {
+        return _timeline.HasEvent(time);
+    }
 
-        public Point<Time, List<Transaction>> MostRecent(Time time)
-        {
-            return _timeline.MostRecent(time);
-        }
+    public IEnumerator<ITimeline<Time, List<Transaction>>.Point> GetEnumerator()
+    {
+        return _timeline.GetEnumerator();
+    }
 
-        public Point<Time, List<Transaction>> First()
-        {
-            return _timeline.First();
-        }
+    public ITimeline<Time, List<Transaction>>.Point MostRecent(Time time)
+    {
+        return _timeline.MostRecent(time);
+    }
+
+    public ITimeline<Time, List<Transaction>>.Point First()
+    {
+        return _timeline.First();
     }
 }
